Implemented mixed precision feature
- The main idea: halve memory bandwidth by only storing f32 instead of f64
- Multi-threading currently overflows L3 cache; mixed precision effectively doubles the throughput
- Some tests still need to be rewritten, but calculation times are all halved now!

+++ crates/bulk-driver/src/channel.rs
@@ -67,7 +67,7 @@ impl CompactBandResultExt for CompactBandResult {
                 let eigenvectors: Vec<Vec<[f64; 2]>> = ea_result.eigenvectors.iter()
                     .map(|field| {
                         field.as_slice().iter()
-                            .map(|c| [c.re, c.im])
+                            .map(|c| [c.re as f64, c.im as f64])
                             .collect()
                     })
                     .collect();
